fixed "jumpy group" bug
CSS was resizing groups on dragover to show the border, and causing
groups to jump away from the card being dragged in as the group was
moved down (it no longer fit in-line due to its new dimensions).

I offset the hidden borders with padding, which is then removed on
dragover.